Revision

+++ 0x01-variables_if_else_while/6-print_numberz.c
@@ -1,18 +1,13 @@
-#include <stdlib.h>
-#include <time.h>
 #include <stdio.h>
-/* more headers goes there */
-
-/* betty style doc for function main goes there */
 
 /**
-* main - Entry point of program
+* main - print all digits in base 10 without using char
 *
 * Return: 0 upon successful execution
 */
 int main(void)
 {
-	int digit = 48;
+	short digit = 48;
 
 	while (digit <= 57)
 		putchar(digit++);
